split bsDssTable and QTable
